requirements.txt
torch
transformers
datasets
scikit-learn
numpy
tqdm
streamlit
gradio
matplotlib
pandas
python-dotenv